refactor swoole_mysql.
