+++ package.xml
@@ -105,6 +105,7 @@
 			<file role="src" name="swoole_websocket_server.c" />
 			<file role="src" name="swoole_atomic.c" />
 			<file role="src" name="swoole_mysql.c" />
+			<file role="src" name="swoole_mysql.h" />
 			<file role="src" name="swoole_redis.c" />
 			<dir name="thirdparty">
 				<file role="src" name="php_http_parser.c" />
